Done a lot of things
- Modified commands #init to ask for a commands array
- Continued to implement new database structure
- Implement guildCreate event to add the new guild to the database
- Moved the registerCommands function to util and added a call to it at the start of the program
- Moved postgres local connection string to .env
- Started implementing config command:
  - Added json for the command
  - Wrote basic structure of the command
  - Implemented #channelsGet function

+++ src/events/guildMemberRemove.ts
@@ -7,7 +7,8 @@ async function execute(member: GuildMember, data: Data) {
 
 	if (data.bumps.get(member.user.id)! <= 0) return;
 
-	const message = (await (data.config!.get("bumpCountChannel") as TextChannel).messages!.fetch()).filter((m) => m.content.startsWith(`<@${member.user.id}>`)).first()!;
+	const message = [...(await (data.config!.get("bumpCountChannel") as TextChannel).messages!.fetch()).values()].find((m) => m.content.startsWith(`<@${member.user.id}>`))!;
+	if(!message) return;
 
 	await message.delete();
